# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for MyBank Application",
        "title": "MyBank API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/features/click": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Record a user's click on a mobile banking feature",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "features"
                ],
                "summary": "Track feature click",
                "parameters": [
                    {
                        "description": "Feature Click Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_GantangSatria_MyBank-BE_pkg_dto_request.TrackFeatureClickRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_GantangSatria_MyBank-BE_pkg_dto_response.Base"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_GantangSatria_MyBank-BE_pkg_dto_response.Base"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_GantangSatria_MyBank-BE_pkg_dto_response.Base"
                        }
                    }
                }
            }
        },
        "/features/clicks": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a list of all features clicked by the user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "features"
                ],
                "summary": "Get feature clicks",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_GantangSatria_MyBank-BE_pkg_dto_response.Base"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_GantangSatria_MyBank-BE_pkg_dto_response.FeatureClickResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_GantangSatria_MyBank-BE_pkg_dto_response.Base"
                        }
                    }
                }
            }
        },
        "/recommendations": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get personalized recommendations for the logged-in user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recommendations"
                ],
                "summary": "Get user recommendations",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_GantangSatria_MyBank-BE_pkg_dto_response.Base"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_GantangSatria_MyBank-BE_pkg_dto_response.RecommendationResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_GantangSatria_MyBank-BE_pkg_dto_response.Base"
                        }
                    }
                }
            }
        },
        "/recommendations/{id}/click": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Record a click on a specific recommendation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recommendations"
                ],
                "summary": "Track recommendation click",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Recommendation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_GantangSatria_MyBank-BE_pkg_dto_response.Base"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_GantangSatria_MyBank-BE_pkg_dto_response.Base"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_GantangSatria_MyBank-BE_pkg_dto_response.Base"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_GantangSatria_MyBank-BE_pkg_dto_response.Base"
                        }
                    }
                }
            }
        },
        "/recommendations/{id}/reason": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get the reason why a specific recommendation was shown to the user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recommendations"
                ],
                "summary": "Get recommendation reason",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Recommendation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_GantangSatria_MyBank-BE_pkg_dto_response.Base"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_GantangSatria_MyBank-BE_pkg_dto_response.RecommendationReasonResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_GantangSatria_MyBank-BE_pkg_dto_response.Base"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_GantangSatria_MyBank-BE_pkg_dto_response.Base"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_GantangSatria_MyBank-BE_pkg_dto_response.Base"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_GantangSatria_MyBank-BE_pkg_dto_request.TrackFeatureClickRequest": {
            "type": "object",
            "required": [
                "feature_name"
            ],
            "properties": {
                "feature_name": {
                    "type": "string",
                    "maxLength": 100
                }
            }
        },
...
        "github_com_GantangSatria_MyBank-BE_pkg_dto_response.Base": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "meta": {
                    "$ref": "#/definitions/github_com_GantangSatria_MyBank-BE_pkg_dto_response.Meta"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "github_com_GantangSatria_MyBank-BE_pkg_dto_response.FeatureClickResponse": {
            "type": "object",
            "properties": {
                "click_count": {
                    "type": "integer"
                },
                "feature_name": {
                    "type": "string"
                },
                "last_clicked": {
                    "type": "string"
                }
            }
        },
...
        "github_com_GantangSatria_MyBank-BE_pkg_dto_response.Meta": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total_items": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "github_com_GantangSatria_MyBank-BE_pkg_dto_response.RecommendationReasonResponse": {
            "type": "object",
            "properties": {
                "consent_status": {
                    "type": "boolean"
                },
                "data_used": {
                    "description": "e.g. \"Riwayat transaksi, Preferensi kategori\"",
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "recommendation_id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "github_com_GantangSatria_MyBank-BE_pkg_dto_response.RecommendationResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "priority": {
                    "type": "integer"
                },
                "reason": {
                    "description": "explainable: \"Karena Anda sering bertransaksi di E-Wallet\"",
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}